interpolated data becomes a float.

+++ thesis_code/dataloading/transforms.py
@@ -34,12 +34,16 @@ class Resize(MRITransform):
         if image.shape[1:] == self.size:
             return sample
 
-        resized_image = F.interpolate(
-            image.unsqueeze(0),  # Add batch and channel dimensions
-            size=self.size,
-            mode="trilinear",
-            align_corners=True,
-        ).squeeze(0)  # Remove batch dimension
+        resized_image = (
+            F.interpolate(
+                image.unsqueeze(0),  # Add batch and channel dimensions
+                size=self.size,
+                mode="trilinear",
+                align_corners=True,
+            )
+            .squeeze(0)
+            .float()
+        )  # Remove batch dimension
         sample["image"] = resized_image  # Remove batch and channel dimensions
         return sample
